encapsulate shopper context code in useShopperContextSearchParams hook
