+++ packages/template-retail-react-app/app/hooks/use-shopper-context-search-params.js
@@ -5,7 +5,11 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import {useEffect} from 'react'
 import {useLocation} from 'react-router-dom'
+import {useUsid, useShopperContext, useShopperContextsMutation} from '@salesforce/commerce-sdk-react'
+import {isServer, isHydrated} from '@salesforce/retail-react-app/app/utils/utils'
+import {useQueryClient} from '@tanstack/react-query'
 
 // Constants
 import {
@@ -16,17 +20,64 @@ import {
 } from '@salesforce/retail-react-app/app/constants'
 
 /*
- * This hook will return all the location search params pertinant
- * to shopper context.
+ * This hook will set the shopper context when search params pertinant
+ * to shopper context are present.
  */
-export const useShopperContextSearchParams = (arraySearchParams = SHOPPER_CONTEXT_ARRAY_FIELDS) => {
+export const useShopperContextSearchParams = (siteId, arraySearchParams = SHOPPER_CONTEXT_ARRAY_FIELDS) => {
+    const {usid} = useUsid()
+    const queryClient = useQueryClient()
     const {search} = useLocation()
     const searchParamsObj = new URLSearchParams(search)
+    const createShopperContext = useShopperContextsMutation('createShopperContext')
+    const updateShopperContext = useShopperContextsMutation('updateShopperContext')
+    const {data: shopperContext} = useShopperContext(
+        {
+            parameters: {usid, siteId}
+        },
+        {
+            enabled: !isServer,
+            onError: async () => {
+                await createShopperContext.mutateAsync({
+                    parameters: {usid, siteId},
+                    body: {}
+                })
+            }
+        }
+    )
+    console.log('shopperContext', shopperContext)
+    const updatedShopperContext = getShopperContextSearchParams(searchParamsObj, arraySearchParams)
+    console.log('updatedshopperContext', updatedShopperContext)
+    const refetchDataOnClient = () => {
+        queryClient.invalidateQueries()
+    }
 
+    useEffect(() => {
+        const executeShopperContextUpdate = async (usid, siteId, updatedShopperContext) => {
+            // update the shopper context if the query string contains the relevant search parameters
+            await updateShopperContext.mutateAsync({
+                parameters: {usid, siteId: siteId},
+                body: updatedShopperContext
+            })
+            // Refresh to update the data on the page
+            refetchDataOnClient()
+        }
+        executeShopperContextUpdate(usid, siteId, updatedShopperContext)
+    }, [search])
+
+    useEffect(() => {
+        console.log('isHydrated()', isHydrated())
+        if (isHydrated()) {
+            refetchDataOnClient()
+        }
+    }, [])
+}
+
+export const getShopperContextSearchParams = (searchParamsObj, arraySearchParams = SHOPPER_CONTEXT_ARRAY_FIELDS) => {
     const getSearchParam = (key, isList = false) =>
         searchParamsObj.has(key)
             ? {[key]: isList ? searchParamsObj.getAll(key) : searchParamsObj.get(key)}
             : {}
+
     const reduceParams = (keys) =>
         keys.reduce(
             (acc, key) => ({...acc, ...getSearchParam(key, arraySearchParams.includes(key))}),
@@ -39,10 +90,10 @@ export const useShopperContextSearchParams = (arraySearchParams = SHOPPER_CONTEX
     const assignmentQualifiers = reduceParams(
         Object.values(SHOPPER_CONTEXT_ASSIGNMENT_QUALIFIERS_SEARCH_PARAMS)
     )
-    const shopperContextObj = reduceParams(Object.values(SHOPPER_CONTEXT_SEARCH_PARAMS))
+    const shopperContextSearchParams = reduceParams(Object.values(SHOPPER_CONTEXT_SEARCH_PARAMS))
 
     return {
-        ...shopperContextObj,
+        ...shopperContextSearchParams,
         ...(Object.keys(customQualifiers).length && {customQualifiers}),
         ...(Object.keys(assignmentQualifiers).length && {assignmentQualifiers})
     }
